myproject/cours/serializers.py
from rest_framework import serializers
from .models import Course
from .models import StudentCourse
class CourseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Course
        fields = '__all__'


class StudentCourseSerializer(serializers.ModelSerializer):
    class Meta:
        model = StudentCourse
        fields = '__all__'